+++ pkg/chainapi/handlers/autorelay.go
@@ -3,8 +3,8 @@ package handlers
 import (
 	"github.com/libp2p/go-libp2p-core/peer"
 	maddr "github.com/multiformats/go-multiaddr"
+	"github.com/rumsystem/quorum/internal/pkg/conn/p2p"
 	"github.com/rumsystem/quorum/internal/pkg/nodectx"
-	"github.com/rumsystem/quorum/internal/pkg/p2p"
 )
 
 type AddRelayParam []string
